product_2/pyproject.toml
[tool.poetry]
name = "product_2"
version = "0.1.0"
description = "A FastAPI-based service for managing products, consuming messages, and performing DB operations"
authors = ["Saqib Murtaza <saqibmurtazakhan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
psycopg = {extras = ["binary"], version = "^3.1.19"}
aiokafka = "^0.10.0"
sqlmodel = "^0.0.19"
kafka = "^1.3.5"
protobuf = "^5.27.1"
uvicorn = "^0.30.1"
requests = "^2.32.3"
dapr = "^1.13.0"
pydantic-settings = "^2.4.0"
supabase = "^2.6.0"
pyjwt = "^2.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
